# doc/openapi.yaml
swagger: '2.0'
info:
  contact: {}
  title: Breezy API
  description: API agrégée de la plateforme Breezy (réseau social microservices).
  version: '1.0'
paths:
...
  /posts/me/liked-ids:
    get:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - likes
      summary: IDs des posts que j'ai likés
      responses:
        '200':
          description: Liste d'IDs
          schema:
            type: array
            items:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
...
  /posts/{id}/comments:
    get:
      produces:
      - application/json
      tags:
      - comments
      summary: Commentaires d'un post
      parameters:
      - type: string
        description: Post ID
        name: id
        in: path
        required: true
      - type: integer
        description: Nb résultats
        name: limit
        in: query
      - type: integer
        description: Décalage
        name: offset
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.Comment'
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
    post:
      security:
      - BearerAuth: []
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - comments
      summary: Commenter un post
      parameters:
      - type: string
        description: Post ID
        name: id
        in: path
        required: true
      - description: Contenu + parent_id optionnel
        name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateCommentRequest'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        '400':
          description: Bad Request
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
  /posts/{id}/comments/{commentId}:
    delete:
      security:
      - BearerAuth: []
      tags:
      - comments
      summary: Supprimer un commentaire (auteur/modérateur/admin)
      parameters:
      - type: string
        description: Post ID
        name: id
        in: path
        required: true
      - type: string
        description: Comment ID
        name: commentId
        in: path
        required: true
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '403':
          description: Forbidden
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
  /posts/{id}/comments/{commentId}/replies:
    get:
      produces:
      - application/json
      tags:
      - comments
      summary: Réponses à un commentaire
      parameters:
      - type: string
        description: Post ID
        name: id
        in: path
        required: true
      - type: string
        description: Comment ID
        name: commentId
        in: path
        required: true
      - type: integer
        description: Nb résultats
        name: limit
        in: query
      - type: integer
        description: Décalage
        name: offset
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.Comment'
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
  /posts/{id}/like:
    post:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - likes
      summary: Liker un post
      parameters:
      - type: string
        description: Post ID
        name: id
        in: path
        required: true
      responses:
        '200':
          description: likes_count
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
    delete:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - likes
      summary: Retirer son like
      parameters:
      - type: string
        description: Post ID
        name: id
        in: path
        required: true
      responses:
        '200':
          description: likes_count
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
  /posts/{id}/likes:
    get:
      produces:
      - application/json
      tags:
      - likes
      summary: Utilisateurs ayant liké un post
      parameters:
      - type: string
        description: Post ID
        name: id
        in: path
        required: true
      responses:
        '200':
          description: Liste d'user IDs
          schema:
            type: array
            items:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
...
definitions:
...
  models.Comment:
    type: object
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      parent_id:
        type: string
      post_id:
        type: string
      reply_count:
        type: integer
      updated_at:
        type: string
...
  models.CreateCommentRequest:
    type: object
    required:
    - content
    properties:
      content:
        type: string
        maxLength: 280
      parent_id:
        type: string
...
host: localhost:8080
basePath: /